feat: permissonamento por id

+++ src/main/java/com/ifoodWebBackEnd/services/FoodService.java
@@ -7,7 +7,10 @@ import com.ifoodWebBackEnd.dtos.FoodResponseDTO;
 import com.ifoodWebBackEnd.repositories.FoodRepository;
 import com.ifoodWebBackEnd.repositories.RestaurantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -24,25 +27,37 @@ public class FoodService {
         return repository.findAll().stream().map(FoodResponseDTO::new).toList();
     }
 
-    public FoodResponseDTO saveFood(FoodRequestDTO data, Long updateUser){
-        Food food = new Food(data, restaurantRepository.findRestaurantById(data.restaurantId()), userService.findUserById(updateUser));
+    public FoodResponseDTO saveFood(FoodRequestDTO data, JwtAuthenticationToken token){
+        Long updateUser = Long.parseLong(token.getName());
+        Restaurant restaurant = restaurantRepository.findRestaurantById(updateUser);
+        if(restaurant == null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+        Food food = new Food(data, restaurant, userService.findUserById(updateUser));
         repository.save(food);
         return new FoodResponseDTO(food);
     }
 
-    public FoodResponseDTO updateFood(Long id, FoodRequestDTO data, Long updateUser){
+    public FoodResponseDTO updateFood(Long id, FoodRequestDTO data, JwtAuthenticationToken token){
+        Long updateUser = Long.parseLong(token.getName());
         Food food = repository.getReferenceById(id);
+        if(updateUser != food.getRestaurant().getId() && token.getToken().getClaimAsString("scope") == "RESTAURANT"){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         food.setName(data.name());
         food.setImage(data.image());
         food.setPrice(data.price());
-        food.setRestaurant(restaurantRepository.getReferenceById(data.restaurantId()));
-        food.setUser(userService.findUserById(updateUser));
+        food.setUpdateUser(userService.findUserById(updateUser));
         return new FoodResponseDTO(food);
     }
 
-    public void deleteFood(Long id, Long updateUser){
+    public void deleteFood(Long id, JwtAuthenticationToken token){
+        Long updateUser = Long.parseLong(token.getName());
         Food food = repository.getReferenceById(id);
-        food.setUser(userService.findUserById(updateUser));
+        food.setUpdateUser(userService.findUserById(updateUser));
+        if(updateUser != food.getRestaurant().getId() && token.getToken().getClaimAsString("scope") == "RESTAURANT"){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         repository.deleteById(id);
     }
 }
